requirements.txt
aiofiles==22.1.0
anyio==3.6.1
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.79.0
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
passlib==1.7.4
protobuf==3.20.1
pyasn1==0.4.8
pydantic==1.9.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.4.0
uvicorn==0.18.2